learning/requirements.txt
numpy==1.21.0
scipy==1.7.1
scikit-learn==0.24.2
absl-py==0.13.0
